Extend Ethereum Gateways to mint ERC20/721 on demand to honor withdrawals (#128)
When the Ethereum Gateway doesn't have a sufficient amount of tokens to
honor a withdrawal it will attempt to mint just enough ERC20 tokens to
complete the withdrawal.

In the case of an ERC721 withdrawal the Ethereum Gateway will first assume
it owns the token being withdrawn and will attempt to transfer it to the
withdrawer, if that fails it attempts to mint it directly to the withdrawer.

In order for on-demand minting to work the Ethereum token contracts must
implement a public `mintTo` function that allows the Gateway to mint tokens.